refactor: 컴포넌트명 컨벤션에 맞춰 수정 - 북마크, 좋아요, 댓글 아이콘 변경 - 사용하지 않는 아이콘 삭제
