+++ src/components/home/card/buttons/BookmarkButton.jsx
@@ -1,8 +1,6 @@
 import React from "react";
 import styled from "styled-components";
 import PropTypes from "prop-types";
-import EmptyBookmarkIcon from "../../../../css/icon/EmptyBookmarkIcon.svg";
-import FullBookmarkIcon from "../../../../css/icon/FullBookmarkIcon.svg";
 
 const Wrapper = styled.span`
   display: flex;
@@ -10,17 +8,27 @@ const Wrapper = styled.span`
   align-items: center;
   margin-right: 12px;
 `;
+
 const Icon = styled.img`
   margin-right: 2px;
 `;
+
 function BookmarkBtn({ bookmarkNumber, myBookmark }) {
   return (
     <Wrapper>
-      <Icon src={myBookmark ? EmptyBookmarkIcon : FullBookmarkIcon} alt="img" />
+      <Icon
+        src={
+          myBookmark
+            ? `${process.env.PUBLIC_URL}/img/bookmark.svg`
+            : `${process.env.PUBLIC_URL}/img/bookmarked.svg`
+        }
+        alt="북마크"
+      />
       <span>{bookmarkNumber}</span>
     </Wrapper>
   );
 }
+
 BookmarkBtn.propTypes = {
   myBookmark: PropTypes.bool,
   bookmarkNumber: PropTypes.number,
