+++ src/components/home/card/buttons/index.js
@@ -1,4 +1,3 @@
-export { default as BookmarkBtn } from "./BookmarkBtn";
-export { default as CommentBtn } from "./CommentBtn";
-export { default as HeartBtn } from "./HeartBtn";
-export { default as Watched } from "./Watched";
+export { default as BookmarkButton } from "./BookmarkButton";
+export { default as CommentButton } from "./CommentButton";
+export { default as LikeButton } from "./LikeButton";
